Packaged the files into packages
