+++ src/main/Data/Duke.txt
@@ -1,3 +1,3 @@
-todo pending make some cash
-event done do more stuff
-12th of December 1999, 7.25pm
+todo done make some money
+event done make some cash /at 12/12/1999 1800
+12th of December 1999, 6pm
