+++ src/main/java/duke/command/AddTodoCommand.java
@@ -1,3 +1,11 @@
+package duke.command;
+
+import duke.tasklist.TaskList;
+import duke.storage.Storage;
+import duke.ui.UI;
+import duke.exception.DukeException;
+import duke.task.Todo;
+
 public class AddTodoCommand extends Command {
 
     public AddTodoCommand(String message) {
