+++ src/main/java/duke/command/Command.java
@@ -1,3 +1,9 @@
+package duke.command;
+
+import duke.tasklist.TaskList;
+import duke.storage.Storage;
+import duke.ui.UI;
+
 public abstract class Command {
     protected String input;
     public Command(String message) {
