+++ src/main/java/duke/command/DoneCommand.java
@@ -1,3 +1,10 @@
+package duke.command;
+
+import duke.tasklist.TaskList;
+import duke.storage.Storage;
+import duke.ui.UI;
+import duke.exception.DukeException;
+
 public class DoneCommand extends Command {
     private int index;
     public DoneCommand(String message) {
