+++ src/main/java/duke/command/ErrorCommand.java
@@ -1,3 +1,10 @@
+package duke.command;
+
+import duke.tasklist.TaskList;
+import duke.storage.Storage;
+import duke.ui.UI;
+import duke.exception.DukeException;
+
 public class ErrorCommand extends Command {
     private UI errorPrinter;
     public ErrorCommand(String message) {
