+++ src/main/java/duke/task/Deadline.java
@@ -1,8 +1,11 @@
+package duke.task;
+
 public class Deadline extends Task {
 
     public Deadline(String description, String extraInfo) {
         super(description);
         this.extraInfo = extraInfo;
+        this.type = "deadline";
     }
 
     @Override
