+++ src/main/java/duke/task/Todo.java
@@ -1,6 +1,9 @@
+package duke.task;
+
 public class Todo extends Task {
     public Todo(String description) {
         super(description);
+        this.type = "todo";
     }
 
     @Override
