+++ src/main/java/duke/ui/UI.java
@@ -1,7 +1,11 @@
-import java.util.ArrayList;
+package duke.ui;
+
 import java.io.BufferedReader;
 import java.io.InputStreamReader;
 import java.io.IOException;
+import duke.tasklist.TaskList;
+import duke.task.Task;
+import duke.exception.DukeException;
 
 public class UI {
